Not calling api calls of announcement and closed user from eReferrals variant.
